extract commands from __main__
so commands can be used in other scripts
